requirements.txt
langchain
langchain-community
openai
faiss-cpu
python-dotenv
streamlit
python-docx
PyMuPDF
